Fix orientation of the drivers for AK8963, AK8975 and MPU9250 MAGs
Due to the unsual axis definition of the AKxxxx MAGs (the +Y axis
is 90ºCW from the +X axis when +Z points up), the driver needs to
invert e.g. X and Z, then rotate rather than inverting all of
X, Y and Z like it was doing.

It seems this bug was introduced in 3d024c6f3 back in November,
between 1.8 and 1.9 and some targets were written or updated in the
meantime assuming the driver was correct and thus have incorrect mag
alignments. I've tried to fix them all, but I don't have all these
boards and while ones with the MP9250 are obvious (assumming the
acc/gyro orientation is right, the mag orientation can be derived
from it), some of them have AK8963 or AK8975 which might be
oriented differently than the acc/gyro.

Fixes #3438
Fixes #3434